+++ realm/realm-library/src/main/java/io/realm/internal/fields/CachedFieldDescriptor.java
@@ -86,7 +86,6 @@ class CachedFieldDescriptor extends FieldDescriptor {
                 verifyInternalColumnType(currentClassName, currentColumnName, currentColumnType);
                 currentClassName = details.linkedClassName;
             }
-
             columnIndices[i] = details.columnIndex;
             tableNativePointers[i] = (currentColumnType != RealmFieldType.LINKING_OBJECTS)
                     ? NativeObject.NULLPTR